Regra para não excluir especialidade.

+++ src/main/java/com/thiago/barroso/clinica/controller/MedicoController.java
@@ -63,8 +63,13 @@ public class MedicoController {
 	@GetMapping({"/id/{idMed}/excluir/especializacao/{idEsp}"})
 	public String excluirEspecialidadePorMedico(@PathVariable("idMed") Long idMed,
 						 @PathVariable("idEsp") Long idEsp, RedirectAttributes attr) {
-		service.excluirEspecialidadesPorMedico(idMed, idEsp);
-		attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso.");
+		
+		if(service.existeEspecialidadeAgendada(idMed, idEsp)) {
+			attr.addFlashAttribute("falha", "Tem consultas agendadas, exclusão negada.");
+		}else {
+			service.excluirEspecialidadesPorMedico(idMed, idEsp);
+			attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso.");
+		}
 		return "redirect:/medicos/dados";
 	}
